refactor: 프로젝트 회원 조회 기능 클래스 분리
